add prims without rules

+++ src/main/java/models/Edge.java
@@ -7,13 +7,13 @@ import java.util.Objects;
 
 public class Edge {
     final List<Vertice> vertices = new ArrayList<>(); // Should never be bigger than 2
-    final double weight;
+    final int weight;
 
     /**
      * Default Constructor
      * @param weight Value needed to run the edge
      */
-    Edge(Double weight) {this.weight = weight;}
+    Edge(Integer weight) {this.weight = weight;}
 
     /**
      * Creates edge from values
@@ -22,7 +22,7 @@ public class Edge {
      * @param v1     first vertice
      * @param v2     second vertice
      */
-    Edge(double weight, int v1, int v2) {
+    Edge(int weight, int v1, int v2) {
         this.vertices.add(new Vertice(v1));
         this.vertices.add(new Vertice(v2));
         this.weight = weight;
@@ -49,7 +49,7 @@ public class Edge {
         String[] strings = fileString.split(";");
         this.vertices.add(new Vertice(Integer.parseInt(strings[0].strip())));
         this.vertices.add(new Vertice(Integer.parseInt(strings[1].strip())));
-        this.weight = Double.parseDouble(strings[2]);
+        this.weight = Integer.parseInt(strings[2]);
     }
 
     /**
